+++ lib/src/dialogs/dialogs.dart
@@ -1,0 +1,8 @@
+library dialogs;
+
+import 'package:flutter/material.dart';
+import 'package:layrz_theme/layrz_theme.dart';
+import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+
+part 'src/general.dart';
+part 'src/action.dart';